+++ packages/app/src/app/overmind/namespaces/profile/state.ts
@@ -5,7 +5,7 @@ import {
 } from 'app/graphql/types';
 import { RootState } from 'app/overmind';
 import { derived } from 'overmind';
-import { SandboxType } from 'app/pages/Profile2/constants';
+import { SandboxType } from 'app/pages/Profile/constants';
 
 export type ProfileCollection = Pick<
   Collection,
