+++ packages/app/src/app/pages/Profile/AllSandboxes.tsx
@@ -308,4 +308,3 @@ const UpgradeBanner = () => {
     </Stack>
   );
 };
-
