+++ packages/app/src/app/pages/common/Modals/MinimumPrivacyModal/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MinimumPrivacy } from 'app/pages/Profile2/MinimumPrivacy';
+import { MinimumPrivacy } from 'app/pages/Profile/MinimumPrivacy';
 
 export const MinimumPrivacyModal: React.FC = props => (
   <MinimumPrivacy {...props} />
