+++ packages/app/src/app/pages/common/Modals/SandboxPickerModal/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SandboxPicker } from 'app/pages/Profile2/SandboxPicker';
+import { SandboxPicker } from 'app/pages/Profile/SandboxPicker';
 
 export const SandboxPickerModal: React.FC = props => (
   <SandboxPicker {...props} />
