+++ packages/app/src/app/pages/index.tsx
@@ -58,8 +58,8 @@ const NotFound = Loadable(() =>
     })
   )
 );
-const Profile2 = Loadable(() =>
-  import(/* webpackChunkName: 'page-profile' */ './Profile2').then(module => ({
+const Profile = Loadable(() =>
+  import(/* webpackChunkName: 'page-profile' */ './Profile').then(module => ({
     default: module.Profile,
   }))
 );
@@ -165,8 +165,8 @@ const RoutesComponent: React.FC = () => {
             <Route path="/signin/duplicate" component={DuplicateAccount} />
             <Route path="/signup/:userId" exact component={SignUp} />
             <Route path="/signin/:jwt?" component={SignInAuth} />
-            <Route path="/u/:username" component={Profile2} />
-            <Route path="/u2/:username" component={Profile2} />
+            <Route path="/u/:username" component={Profile} />
+            <Route path="/u2/:username" component={Profile} />
             <Route path="/search" component={Search} />
             <Route path="/patron" component={Patron} />
             <Route path="/pro" component={Pro} />
